cursor-artifacts/requirements_add.txt
cryptography==41.0.7
django-ratelimit==4.1.0
sentry-sdk==1.40.6